Module 5/requirements.txt
fastapi==0.78.0
uvicorn==0.17.0
openai>=1.0.0
pinecone-client==2.1.0
